Depthwise conv kernel
This change contains 2 vectorized dwconv kernel implementation, one
generic version and a specialized version for 3x3 "stride 1" (to be more
precise, stride_w == 1x dilation_w) where 6 of the 9 inputs can be
reused in the next pixel on the same row.

The generic version works on up to 32 channels at a time, not
attempting to preserve input or filter values in vregs.

The "stride 1" version works on up to 16 channels at a time, preserving
all 9 filter vectors and reusing 6 of the 9 input values. To avoid
frequent switching between filter values and quantization params, this
implementation spills the accumulators to be postprocessed later. This
space is currently allocated from the heap - we'll revisit the memory
arrangement later.

Test cases are created with 4x4 output patches.

Benchmark cases are added to replicate dwconv nodes in MobileNet V1.

Change-Id: Iaf65eba593bea009da23ef31a318bc880de7d2c4